# internal/route53dns/model_load_balancer_input.go
/*
 * DNS Service
 *
 * Sevice for managing DNS with route53
 *
 * API version: 0.0.1
 * Contact: iitr.animesh@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package route53dns




type LoadBalancerInput struct {

	Name string `json:"name,omitempty"`

	TagKey string `json:"tag_key,omitempty"`
}

// AssertLoadBalancerInputRequired checks if the required fields are not zero-ed
func AssertLoadBalancerInputRequired(obj LoadBalancerInput) error {
	return nil
}

// AssertLoadBalancerInputConstraints checks if the values respects the defined constraints
func AssertLoadBalancerInputConstraints(obj LoadBalancerInput) error {
	return nil
}